Fix vigra connected components

+++ neurofire/transforms/segmentation.py
@@ -272,7 +272,7 @@ class ConnectedComponents3D(Transform):
             raise NotImplementedError("Connected components is not supported without vigra "
                                       "if label_segmentation is set to True.")
         if self.label_segmentation:
-            connected_components = vigra.analysis.labelVolumeWithBackground(volume)
+            connected_components = vigra.analysis.labelVolumeWithBackground(volume.astype('uint32'))
         else:
             connected_components, _ = label(volume)
         return connected_components
